move guardian set to a const file
